% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_process_result.R
\name{format.kreuzberg_code_process_result}
\alias{format.kreuzberg_code_process_result}
\title{Format method for kreuzberg_code_process_result}
\usage{
\method{format}{kreuzberg_code_process_result}(x, ...)
}
\arguments{
\item{x}{A \code{kreuzberg_code_process_result} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A character string representation.
}
\description{
Format method for kreuzberg_code_process_result
}
