% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_process_result.R
\name{print.kreuzberg_code_process_result}
\alias{print.kreuzberg_code_process_result}
\title{Print method for kreuzberg_code_process_result}
\usage{
\method{print}{kreuzberg_code_process_result}(x, ...)
}
\arguments{
\item{x}{A \code{kreuzberg_code_process_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for kreuzberg_code_process_result
}
